requirements.txt
# Django and core dependencies
Django>=4.2,<5.0
gunicorn>=21.0
psycopg2-binary>=2.9
python-decouple>=3.8
whitenoise>=6.5
dj-database-url>=2.1.0

# Production dependencies
gevent>=23.0

# Security
django-cors-headers>=4.3

# Development and testing (optional)
pytest>=7.4
pytest-django>=4.5
flake8>=6.0

# Monitoring and logging
django-extensions>=3.2
